move tests to own file
