+++ packages/batch-delegate/tests/basic.example.test.ts
@@ -353,160 +353,4 @@ describe('batch delegation within basic stitching example', () => {
       ],
     });
   });
-
-  test('works with merged types and array batching', async () => {
-    const queries = {
-      schema1: [],
-      schema2: [],
-    };
-    const schema1 = makeExecutableSchema({
-      typeDefs: /* GraphQL */ `
-        type Book {
-          id: ID!
-          title: String
-        }
-
-        type Query {
-          book(id: ID!): Book
-        }
-      `,
-      resolvers: {
-        Query: {
-          book: (_obj, _args, _ctx, info) => {
-            addToQueries(info, 'schema1');
-            return { id: '1', title: 'Book 1' };
-          },
-        },
-      },
-    });
-
-    const schema2 = makeExecutableSchema({
-      typeDefs: /* GraphQL */ `
-        type Book {
-          id: ID!
-          isbn: Int
-        }
-
-        type Query {
-          books(id: [ID!]!): [Book]
-        }
-      `,
-      resolvers: {
-        Query: {
-          books: (_obj, _args, _ctx, info) => {
-            addToQueries(info, 'schema2');
-            return [{ id: '1', isbn: 123 }];
-          },
-        },
-      },
-    });
-
-    const addToQueries = (info: any, schema: string) =>
-      info.operation.selectionSet.selections.forEach((s: any) => {
-        s.selectionSet.selections.forEach((s: any) => {
-          queries[schema].push(s.name.value);
-        });
-      });
-
-    const stitchedSchemaWithValuesFromResults = stitchSchemas({
-      subschemas: [
-        {
-          schema: schema1,
-          merge: {
-            Book: {
-              selectionSet: '{ id }',
-              fieldName: 'book',
-            },
-          },
-        },
-        {
-          schema: schema2,
-          merge: {
-            Book: {
-              selectionSet: '{ id }',
-              fieldName: 'books',
-              key: ({ id }) => id,
-              argsFromKeys: ids => ({ id: ids }),
-              valuesFromResults: ({ results }, keys) => {
-                const response = Object.fromEntries(results.map((r: any) => [r.id, r]));
-
-                return keys.map(key => response[key] ?? { id: key });
-              },
-            },
-          },
-        },
-      ],
-      mergeTypes: true,
-    });
-
-    const stitchedSchemaWithoutValuesFromResults = stitchSchemas({
-      subschemas: [
-        {
-          schema: schema1,
-          merge: {
-            Book: {
-              selectionSet: '{ id }',
-              fieldName: 'book',
-            },
-          },
-        },
-        {
-          schema: schema2,
-          merge: {
-            Book: {
-              selectionSet: '{ id }',
-              fieldName: 'books',
-              key: ({ id }) => id,
-              argsFromKeys: ids => ({ id: ids }),
-            },
-          },
-        },
-      ],
-      mergeTypes: true,
-    });
-
-    const query = /* GraphQL */ `
-      query {
-        book(id: "1") {
-          id
-          title
-          isbn
-        }
-      }
-    `;
-
-    const goodResult = await execute({
-      schema: stitchedSchemaWithoutValuesFromResults,
-      document: parse(query),
-    });
-    const badResult = await execute({
-      schema: stitchedSchemaWithValuesFromResults,
-      document: parse(query),
-    });
-
-    if (isIncrementalResult(goodResult)) throw Error('result is incremental');
-    if (isIncrementalResult(badResult)) throw Error('result is incremental');
-
-    // queries from both goodResult and badResult in the arrays
-    expect(queries).toEqual({
-      schema1: ['id', 'title', '__typename', 'id', 'id', 'title', '__typename', 'id'],
-      schema2: ['id', 'isbn', 'id', 'isbn'],
-    });
-
-    expect(goodResult.data).toEqual({
-      book: {
-        id: '1',
-        title: 'Book 1',
-        isbn: 123,
-      },
-    });
-
-    expect(badResult.data).toEqual({
-      book: {
-        id: '1',
-        title: 'Book 1',
-        isbn: 123,
-      },
-    });
-  });
 });
